requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0